% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badges.R
\docType{data}
\name{badges}
\alias{badges}
\title{Badge data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 8 columns.
}
\usage{
badges
}
\description{
Data on the employee badges. The variable `joined` is the approximate days
prior to Helly joining that the other employees joined Lumon. Mark is 730,
(2 years), since in Season 1 Episode 1 (at 40:39) Mark says he joined about
two years ago, Irving is 3,285 (9 years), since a marketing Linkedin post
from Lumon stated he joined 9 years prior.
}
\keyword{datasets}
