% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_MDR.R
\name{Compute_MDR}
\alias{Compute_MDR}
\title{Compute MDR}
\usage{
Compute_MDR(
  x,
  y,
  z,
  n,
  SCDR_model = c("Hill2"),
  design,
  lower = c(1e-04, 1e-04, 1e-04, 1e-04),
  upper = c(10, 10, 1, 1),
  do.plot = FALSE
)
}
\arguments{
\item{x}{sequence of doses first compound}

\item{y}{sequence of doses second compound}

\item{z}{sequence of responses}

\item{n}{number of binomial samples at various dose values}

\item{SCDR_model}{family of the single compound dose-response curve}

\item{design}{any of the below
'single-parallel': single ray parallel to an axis (dose of one of the two compounds is fixed), single compound experiment data not used,
'single-ray': single ray design (ratio of doses of two compounds is fixed), single compound experiment data not used
TODO 'general': fully arbitrary design, all data used including  single compound experiment data
Each of the design above has to contain single-compound experiment data}

\item{lower}{lower limits for the search of parameter}

\item{upper}{upper limits for the search of parameter}

\item{do.plot}{logical}
}
\value{
estimates of doses or concentrations eliciting 50\% max effect and of MDR defined as d50_DA / d50_obs
}
\description{
computing Model Deviation Ratio for data obtained under three types of designs
}
