% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D.curvirostris.mix.R
\docType{data}
\name{D.curvirostris.mix}
\alias{D.curvirostris.mix}
\title{Data extracted from Dallabonna et al. Chemosphere (2014)}
\format{an R object}
\source{
subset of numerical values from Table 1 and 2 in Dallabonna et al. Chemosphere (2014)
}
\usage{
D.curvirostris.mix
}
\description{
Data extracted from Dallabonna et al. Chemosphere (2014)
}
\keyword{datasets}
