% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR_Bliss_Hill2.R
\name{DR_Bliss_Hill2}
\alias{DR_Bliss_Hill2}
\title{Bivariate dose-response function under Bliss independance model}
\usage{
DR_Bliss_Hill2(d1, d2, a1, b1, a2, b2)
}
\arguments{
\item{d1}{dose of first compound: a numeric vector}

\item{d2}{dose of second compound: a numeric vector}

\item{a1}{parameter a in Hill-2 DR fuction 1/(1+(a/x)**b) for first compound}

\item{b1}{parameter b in Hill-2 DR fuction 1/(1+(a/x)**b) for first compound}

\item{a2}{parameter a in Hill-2 DR fuction 1/(1+(a/x)**b) for second compound}

\item{b2}{parameter b Hill-2 DR fuction in 1/(1+(a/x)**b) for second compound}
}
\value{
vector of numerical responses
}
\description{
Compute bivariate dose-response function when each mixture compoent follows a 2-parameter Hill dose-response function and the Bliis independance  model holds.
}
