% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR_Jonker_Hill2.R
\name{DR_JonkerSA_Hill2}
\alias{DR_JonkerSA_Hill2}
\title{Bivariate dose-response function under Jonker S/A model}
\usage{
DR_JonkerSA_Hill2(d1, d2, a1, b1, a2, b2, alpha)
}
\arguments{
\item{d1}{dose of first compound: a numeric vector}

\item{d2}{dose of second compound: a numeric vector}

\item{a1}{parameter a in Hill-2 DR fuction 1/(1+(a/x)**b) for first compound}

\item{b1}{parameter b in Hill-2 DR fuction 1/(1+(a/x)**b) for first compound}

\item{a2}{parameter a in Hill-2 DR fuction 1/(1+(a/x)**b) for second compound}

\item{b2}{parameter b Hill-2 DR fuction in 1/(1+(a/x)**b) for second compound}

\item{alpha}{parameter quantifying departure from Loewe additivity in Jonker S/A model}
}
\value{
vector of numerical responses
}
\description{
Compute bivariate dose-response function when each mixture compoent follows a 2-parameter Hill dose-response function and Jonker S/A model holds.
}
\references{
Jonker model S/A (Eq. 7 p. 2703 Jonker et al. Env Tox & Chem 24(10) 2005)
}
