% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR_Voelund_Hill2.R
\name{DR_Voelund_Hill2}
\alias{DR_Voelund_Hill2}
\title{Bivariate dose-response function under Voelund model}
\usage{
DR_Voelund_Hill2(d1, d2, a1, b1, a2, b2, lambda1, lambda2)
}
\arguments{
\item{d1}{dose of first compound: a single numeric entry}

\item{d2}{dose of second compound: a single numeric entry}

\item{a1}{parameter a in Hill-2 DR fuction 1/(1+(a/x)**b) for first compound}

\item{b1}{parameter b in Hill-2 DR fuction 1/(1+(a/x)**b) for first compound}

\item{a2}{parameter a in Hill-2 DR fuction 1/(1+(a/x)**b) for second compound}

\item{b2}{parameter b Hill-2 DR fuction in 1/(1+(a/x)**b) for second compound}

\item{lambda1}{parameter Voelund model}

\item{lambda2}{parameter Voelund model}
}
\value{
response: a single numerical value
}
\description{
Compute bivariate dose-response function when each mixture compoent follows a 2-parameter Hill dose-response function and the Voelund model holds.
}
\references{
Eq. 6  p. 388 in Soerensen et al. Environ. Ecol. Stat. 2007
}
