% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevator_dings.R
\docType{data}
\name{elevator_dings}
\alias{elevator_dings}
\title{Elevator Ding Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 91 rows and 8 columns.
}
\source{
https://www.reddit.com/r/SeveranceAppleTVPlus/comments/w8imx3/a_complete_analysis_of_all_of_the_ding_sounds/
}
\usage{
elevator_dings
}
\description{
The first part of this data set was based on data compiled by Reddit user
Sam_Badi, the remainder was added by the package author.
}
\keyword{datasets}
