% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episodes.R
\docType{data}
\name{episodes}
\alias{episodes}
\title{Severance Episode Data}
\format{
## `episodes`
A data frame with 19 rows and 7 columns:
\describe{
  \item{season}{Season}
  \item{episode}{Episode number}
  \item{title}{Title}
  \item{duration}{Duration}
  \item{release_date}{Release Date}
  \item{writer}{Writer}
  \item{director}{Director}
}
}
\source{
https://severance.wiki/list_of_severance_episodes Accessed Jan 26 2025
}
\usage{
episodes
}
\description{
Severance Episode Data
}
\keyword{datasets}
