% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDR.R
\name{FindSampleSize}
\alias{FindSampleSize}
\title{Find the minimized sample size and decision rules}
\usage{
FindSampleSize(
  J,
  phi_TL,
  phi_TH,
  phi_EL,
  phi_EH,
  alpha = 0.1,
  beta = 0.8,
  delta = 0.2,
  Nmin,
  Nmax,
  if.verbose = FALSE
)
}
\arguments{
\item{J}{Number of doses considered in the trial}

\item{phi_TL}{True tolerable toxicity rate for each dose}

\item{phi_TH}{True insuffurable toxicity rate for each dose}

\item{phi_EL}{True insignificant efficacy rate for each dose}

\item{phi_EH}{True desirable efficacy rate for each dose}

\item{alpha}{Targeted type I error rate}

\item{beta}{Targeted general power}

\item{delta}{Targeted plateau criteria}

\item{Nmin}{Minimal number of patients considered}

\item{Nmax}{Maximum number of patients considered}

\item{if.verbose}{Display extended information}
}
\value{
A list contains the estimated design parameters.
\itemize{
    \item best_N - Estimated minimal sample size.
    \item best_mT - Estimated decision rule for toxicity endpoints (counts).
    \item best_mTr - Estimated decision rule for toxicity endpoints (rate).
    \item best_mE - Estimated decision rule for efficacy endpoints (counts).
    \item best_mEr - Estimated decision rule for efficacy endpoints (rate).
    \item best_Ind - Index of the estimated design parameter in Metadata.
    \item meta.data - Metadata for the estimation.
}
}
\description{
This function is designed to estimate the minimized sample size
and decision boundaries of toxicity and efficacy endpoints for trial
enrollment and decision-making.
}
\examples{
res <- FindSampleSize(J = 3, delta = 0.2,
                      phi_TL = 0.2, phi_TH = 0.4,
                      phi_EL = 0.2, phi_EH = 0.4,
                      alpha = 0.1, beta = 0.8,
                      Nmin = 10, Nmax = 100, if.verbose = FALSE)

}
