% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_Single_Hill.R
\name{Fit_Single_Hill}
\alias{Fit_Single_Hill}
\title{Estimate parameters in   Hill model families}
\usage{
Fit_Single_Hill(
  x,
  y,
  n,
  models = c("Hill2", "Hill3", "Hill4"),
  lower = c(1e-04, 1e-04, 1e-04, 1e-04),
  upper = c(10, 10, 1, 1),
  bootstrap = FALSE,
  B = 100,
  select = FALSE,
  eps = 1e-15
)
}
\arguments{
\item{x}{sequence of dose input}

\item{y}{sequence of responses}

\item{n}{number of binomial samples at various dose values}

\item{models}{one of 2-4 parameter Hill function}

\item{lower}{lower limits for the search of parameter}

\item{upper}{upper limits for the search of parameter}

\item{bootstrap}{logical resampling for estimating parameter variability}

\item{B}{number of bootstrap resampling}

\item{select}{TRUE/FALSE perform model selection}

\item{eps}{used to replace 0's by small numerical values}
}
\value{
Estimate of Hill model(s) parameters and bootstrap sample for this estimate
Perform model selection by cross-validation
Return estimate of 50% effect doses and bootstrap sample for this estimate
NB:  50% effect dose may not exist under Hill4 for some combinations (c,d)

Detail
Model assumed: y ~ B(n,f(x))
Hill models:
Hill 4: f(x) = d + (c-d) / ( 1+ (a/x)**b)
Hill 3: d=0
Hill 2: d=0 ; c=1
ML estimation, optimization of log likelihood with Generalized Simulated Annealing (GenSA::GenSA)

0 doses bring numerical issues
0's in x and y replaced by min(x) * 1e-9 / min(y) * 1e-9

Model selection with leave-"some"-out likelihood cross-validation
"some" = all obs receiving a certain dose level
}
\description{
Estimate parameters in single-compound dose-reponse models under some of the Hill model families (2,3,4 parameters)

Parameters
}
\author{
Gilles Guillot \email{gilles.b.guillot@gmail.com}
}
