% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hill.R
\name{Hill}
\alias{Hill}
\title{Hill dose response function}
\usage{
Hill(x, a, b, c = 1, d = 0)
}
\arguments{
\item{x}{dose (vector)}

\item{a}{a in d + (c-d)/(1+(a/x)**b)  ; 50\% max effect}

\item{b}{b in  d + (c-d)/(1+(a/x)**b)}

\item{c}{c in d + (c-d)/(1+(a/x)**b)}

\item{d}{d in d + (c-d)/(1+(a/x)**b)}
}
\value{
response
}
\description{
computation of 2-4 parameter Hill dose response function
}
