% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hill.R
\name{Hill2_inv}
\alias{Hill2_inv}
\title{reciprocal of 2-paramter Hill function}
\usage{
Hill2_inv(y, a, b)
}
\arguments{
\item{y}{response (numeric vector)}

\item{a}{a a in d + (c-d)/(1+(a/x)**b)  ; 50\% max effect}

\item{b}{b b in  d + (c-d)/(1+(a/x)**b)}
}
\value{
doses elicting effects y
}
\description{
computes doses elicting effects y when dose reponse function is Hill2
}
