% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDR_obs.R
\docType{data}
\name{MDR_obs}
\alias{MDR_obs}
\title{Documentation for MDR_obs}
\format{an R object}
\source{
MDR computed from data in  Dallabonna et al. Chemosphere (2014)
}
\usage{
MDR_obs
}
\description{
Documentation for MDR_obs
}
\keyword{datasets}
