% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Resample_MDR.R
\name{Resample_MDR}
\alias{Resample_MDR}
\title{Perform Monte Carlo simulation of MDR statistic}
\usage{
Resample_MDR(
  x,
  y,
  z = NULL,
  n,
  SCDR_model = "Hill2",
  par_Hill = NULL,
  design,
  lower = c(1e-04, 1e-04, 1e-04, 1e-04),
  upper = c(10, 10, 1, 1),
  n_resamp,
  distribution = "binom",
  verbose = TRUE
)
}
\arguments{
\item{x}{sequence of doses first compound}

\item{y}{sequence of doses second compound}

\item{z}{sequence of responses}

\item{n}{number of binomial samples at various dose values}

\item{SCDR_model}{family of the single compound dose-response curve}

\item{par_Hill}{parameters of the Hill functions.
With par_Hill = (a1,b1,a2,b2)
a1: a in 1/(1+(a/x)**b) for first compound
b1: b in  1/(1+(a/x)**b) for first compound
a2: a in  1/(1+(a/x)**b) for second compound
b2: b in 1/(1+(a/x)**b) for second compound}

\item{design}{any of the below
'single-parallel': single ray parallel to an axis (dose of one of the two compounds is fixed), single compound experiment data not used,
'single-ray': single ray design (ratio of doses of two compounds is fixed), single compound experiment data not used
'general' (TODO): fully arbitrary design, all data used including  single compound experiment data}

\item{lower}{lower limits for the search of parameter}

\item{upper}{upper limits for the search of parameter}

\item{n_resamp}{number of dataset resampled}

\item{distribution}{distribution of response conditional on Hill function value}

\item{verbose}{will print the percentage of computation carried out if TRUE}
}
\value{
a list with an element named MDR_resamp
}
\description{
Perform Monte Carlo simulation of MDR statistic
under generative model and experimental designed known/assumed
for the data
}
\details{
Input data should include design (x,y,n) AND
EITHER vector of response mortality values z
OR estimate of dose repose curves parameters par_Hill
}
