% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2e1_transcript.R
\docType{data}
\name{s2e1_transcript}
\alias{s2e1_transcript}
\title{Season 2 Episode 1 Transcript}
\format{
An object of class \code{data.frame} with 336 rows and 3 columns.
}
\source{
https://severance.wiki/s2e1_transcript Accessed Jan 26 2025
}
\usage{
s2e1_transcript
}
\description{
Season 2 Episode 1 Transcript
}
\keyword{datasets}
