% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2e2_transcript.R
\docType{data}
\name{s2e2_transcript}
\alias{s2e2_transcript}
\title{Severance Episode Data}
\format{
An object of class \code{data.frame} with 298 rows and 3 columns.
}
\source{
https://severance.wiki/s2e2_transcript Accessed Jan 26 2025
}
\usage{
s2e2_transcript
}
\description{
Severance Episode Data
}
\keyword{datasets}
