% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mdr.R
\name{theme_mdr}
\alias{theme_mdr}
\title{Custom MDR ggplot2 Theme}
\usage{
theme_mdr()
}
\value{
A ggplot2 theme object.
}
\description{
Colors are pulled from the MDR consoles in shown in the Severance TV show as
 well as from the (presumable) hex codes in the Lexington Letters.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point() +
  labs(title = "Car Efficiency", subtitle = "MPG vs Weight") +
  theme_mdr()
}
