% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcripts.R
\docType{data}
\name{transcripts}
\alias{transcripts}
\title{Severance Transcripts}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5904 rows and 5 columns.
}
\source{
https://severance.wiki/list_of_severance_episodes Accessed Feb 10 2025
}
\usage{
transcripts
}
\description{
Severance Transcripts
}
\keyword{datasets}
