% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badges.R
\docType{data}
\name{badges}
\alias{badges}
\title{Badge data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 8 columns.
}
\usage{
badges
}
\description{
Badge data
}
\keyword{datasets}
