% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dallabona.mix.R
\docType{data}
\name{dallabona.mix}
\alias{dallabona.mix}
\title{Documentation for dallabona.mix}
\format{an R object}
\source{
subset of numerical values from Table 1 and 2 in Dallabonna et al. Chemosphere (2014)
}
\usage{
dallabona.mix
}
\description{
Documentation for dallabona.mix
}
\keyword{datasets}
