% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dallabona.sc.R
\docType{data}
\name{dallabona.sc}
\alias{dallabona.sc}
\title{Documentation for dallabona.sc}
\format{an R object}
\source{
subset of numerical values from Table 1 and 2 in Dallabonna et al. Chemosphere (2014)
}
\usage{
dallabona.sc
}
\description{
Documentation for dallabona.sc
}
\keyword{datasets}
