% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevator_dings.R
\docType{data}
\name{elevator_dings}
\alias{elevator_dings}
\title{Elevator Ding Data}
\format{
## `elevator_dings`
A data frame with 48 rows and 8 columns:
\describe{
  \item{season}{Season}
  \item{episode}{Episode number}
  \item{time}{Time}
  \item{character}{Character}
  \item{pitch}{Pitch of elevator ding}
  \item{io}{Innie or Outie?}
  \item{action}{Action, one of: enters, sleep, wakes up}
  \item{note}{Notes}
}
}
\source{
https://www.reddit.com/r/SeveranceAppleTVPlus/comments/w8imx3/a_complete_analysis_of_all_of_the_ding_sounds/
}
\usage{
elevator_dings
}
\description{
Elevator Ding Data
}
\keyword{datasets}
