% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDR_sim.R
\docType{data}
\name{MDR_sim}
\alias{MDR_sim}
\title{Documentation for MDR_sim}
\format{an R object}
\source{
MDR resampled from data in  Dallabonna et al. Chemosphere (2014)
}
\usage{
MDR_sim
}
\description{
Documentation for MDR_sim
}
\details{
rows contain compound combinations and concentration ratio, cf row names
}
\keyword{datasets}
