% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_Mixture.R
\name{Sim_Mixture}
\alias{Sim_Mixture}
\title{Simulate toxicological/pharmacological mixture experiment data}
\usage{
Sim_Mixture(
  x,
  y,
  n,
  par_Hill,
  interaction_model,
  par_int = NULL,
  distribution,
  par_dist = 1
)
}
\arguments{
\item{x}{sequence of doses first compound. May contaon zeroes}

\item{y}{sequence of doses second compound. May contaon zeroes}

\item{n}{number of binomial samples at various dose values}

\item{par_Hill}{parameters of the Hill functions.
If par_Hill = (a1,a2,b1,b2)
a1 is a in 1/(1+(a/x)**b) for first compound
b1 is b in  1/(1+(a/x)**b) for first compound
a2 is a in  1/(1+(a/x)**b) for second compound
b2 is b in 1/(1+(a/x)**b) for second compound}

\item{interaction_model}{'Loewe','Bliss','Hewlett','Voelund' or 'JonkerSA'}

\item{par_int}{parameter for interaction model}

\item{distribution}{distribution of response conditional on Hill function value}

\item{par_dist}{param of the previous distribution}
}
\value{
list with input parameters and simulated responses
}
\description{
Simulate toxicological/pharmacological mixture experiment data
}
